refactor: rename data as view
